Refactored standings to use a structured object instead of a DataFrame.
It's becoming tedious to keep various DataFrames in sync. Standings
is a type that I'm hoping to generalize into a simple structured type
that can be converted into a DataFrame. I need to figure out the
metaprogramming magic to configure keys and whatnot in the created
DataFrame.

This was motivated by the need to fetch the current year's standings
from fastf1, but I want to have a way of statically enforcing that
the dataframe structure is consitent between functions.